pyproject.toml
[project]
name = "agent-from-scratch"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.0",
    "pydantic-ai>=0.7.4",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
